CostingSpecification.scala for Context
